- Added up and down attack animations. - Added respawn message. Players needs to press X to spawn. - Fixed anim prio issue. - Fixed crash issue jump anim. - Added 3 ammo shotgun weapon. - Added sword cancel feedback. - Added game state sync when running w/ full client sim. - Made client sim a command line option (clientsim=1). - Added networked player input broadcast. Run client sim on receive. - Added half size top and bottom collision shapes. - Added passthrough block image. - Select spawn pos based on distance to other players.

+++ users/marcel/netarena/gametypes.h
@@ -63,10 +63,10 @@ struct CollisionInfo
 			y1 <= y;
 	}
 
-	int x1;
-	int y1;
-	int x2;
-	int y2;
+	int16_t x1;
+	int16_t y1;
+	int16_t x2;
+	int16_t y2;
 };
 
 struct Pickup
@@ -74,11 +74,11 @@ struct Pickup
 	bool isAlive;
 
 	PickupType type;
-	int blockX;
-	int blockY;
-	int x1;
-	int y1;
-	int x2;
-	int y2;
+	uint8_t blockX;
+	uint8_t blockY;
+	int16_t x1;
+	int16_t y1;
+	int16_t x2;
+	int16_t y2;
 	uint16_t spriteId;
 };
